Created new BaseJDBCIT class to extend BaseJDBCTest to avoid classes that don't need the shared connection field, failing when they don't have the proper connection parameters.
