+++ src/test/java/net/snowflake/client/jdbc/ResultSetMultiTimeZoneLatestIT.java
@@ -29,7 +29,7 @@ import org.junit.runners.Parameterized;
  */
 @RunWith(Parameterized.class)
 @Category(TestCategoryResultSet.class)
-public class ResultSetMultiTimeZoneLatestIT extends BaseJDBCTest {
+public class ResultSetMultiTimeZoneLatestIT extends BaseJDBCIT {
   @Parameterized.Parameters(name = "format={0}, tz={1}")
   public static Collection<Object[]> data() {
     // all tests in this class need to run for both query result formats json and arrow
