+++ src/test/java/net/snowflake/client/jdbc/StatementIT.java
@@ -36,7 +36,7 @@ import org.junit.rules.TemporaryFolder;
 
 /** Statement tests */
 @Category(TestCategoryStatement.class)
-public class StatementIT extends BaseJDBCTest {
+public class StatementIT extends BaseJDBCIT {
   protected static String queryResultFormat = "json";
 
   public static Connection getConnection() throws SQLException {
